services/api/pyproject.toml
[project]
name = "neurosonix-api"
version = "0.1.0"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]",
  "pydantic>=2.8",
  "httpx",
  "psycopg2-binary",
  "sqlalchemy>=2.0",
  "influxdb-client",
  "neo4j",
  "weaviate-client>=4.6.4",
  "python-jose[cryptography]",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "pytest",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
